add android protocol

+++ Lagrange.Core/Internal/Context/HighwayContext.cs
@@ -142,7 +142,7 @@ internal class HighwayContext : ContextBase, IDisposable
             Uin = upBlock.Uin.ToString(),
             Command = "PicUp.DataUp",
             Seq = upBlock.Sequence,
-            AppId = (uint)AppInfo.SubAppId,
+            AppId = AppInfo.SubAppId,
             DataFlag = 16,
             CommandId = (uint)upBlock.CommandId,
         };
